.agents/skills/access-migration-catalog/requirements.txt
# ============================================================
# AI-Compatible Conversion Pipeline — Python dependencies
# ============================================================

# Document conversion
mammoth>=1.8.0           # docx → html (best fidelity Word conversion)
markdownify>=0.13.0      # html → clean markdown
python-pptx>=1.0.0       # pptx/ppt slide text extraction

# Spreadsheet
openpyxl>=3.1.0          # .xlsx (modern Excel)
xlrd>=2.0.1              # .xls (Excel 97-2003 binary only)

# PDF
pymupdf4llm>=0.0.17      # PDF text → LLM-ready markdown; also installs pymupdf (fitz)
# fitz (pymupdf) is used directly for page→PNG rendering — bundled with pymupdf4llm

# RTF
striprtf>=0.0.26         # RTF control code stripping → plain text

# XML / VSDX parsing (stdlib zipfile + lxml, no extra Visio library needed)
lxml>=5.0

# Visio diagram rendering
aspose-diagram-python>=26.0   # .NET-backed; DLLs bundled in wheel; no separate .NET install needed on Windows 10/11

# Access DB via ODBC (requires MS Access Driver installed separately)
pyodbc>=5.0

# Image utilities
Pillow>=10.0

# Catalog charts (already in catalog.py)
matplotlib>=3.8
squarify>=0.4
